refactor: organizing structure of ui module, create new packages, move some files
